restructure factory to utils
