+++ src/development/vortex/development/__init__.py
@@ -1,2 +1,2 @@
 from .version import __version__
-from .core import create_model
+from .utils import create_model
